shader minprecision fixes

+++ WickedEngine/shaders/objectGS_voxelizer.hlsl
@@ -4,18 +4,18 @@
 struct GSInput
 {
 	float4 pos : SV_POSITION;
-	float4 color : COLOR;
-	float4 uvsets : UVSETS;
-	float3 nor : NORMAL;
+	min16float4 color : COLOR;
+	min16float4 uvsets : UVSETS;
+	min16float3 nor : NORMAL;
 };
 
 // Note: centroid interpolation is used to avoid floating voxels in some cases
 struct GSOutput
 {
 	float4 pos : SV_POSITION;
-	centroid float4 color : COLOR;
-	centroid float4 uvsets : UVSETS;
-	centroid float3 N : NORMAL;
+	centroid min16float4 color : COLOR;
+	centroid min16float4 uvsets : UVSETS;
+	centroid min16float3 N : NORMAL;
 	centroid float3 P : POSITION3D;
 
 #ifdef VOXELIZATION_CONSERVATIVE_RASTERIZATION_ENABLED
