+++ WickedEngine/shaders/objectVS_voxelizer.hlsl
@@ -5,9 +5,9 @@
 struct VSOut
 {
 	float4 pos : SV_POSITION;
-	float4 color : COLOR;
-	float4 uvsets : UVSETS;
-	float3 N : NORMAL;
+	min16float4 color : COLOR;
+	min16float4 uvsets : UVSETS;
+	min16float3 N : NORMAL;
 #ifndef VOXELIZATION_GEOMETRY_SHADER_ENABLED
 	float3 P : POSITION3D;
 #endif // VOXELIZATION_GEOMETRY_SHADER_ENABLED
